Refactor path bounds math to helper
